--- .github/workflows/app-review/index.js
@@ -1,5 +1,0 @@
-const license = require('./license');
-
-module.exports = {
-  license,
-};
